68928-spawn-command-isabs-zmedico.patch (spawn included "." in PATH) 98727_sleep_locale_cleanup.patch (sleep was defined twice for "." and "," locales) 98827_lchown_normal_files.patch (lchown was being used on known-to-be-regular files) 98827_darwin_noop_lchown.patch (lchwon was aliased to chown on darwin) 99101_ccache_perms_adjust.patch (perms test and adjustment didn't match) 99120_uid_gid_mixup.patch (portage gid used in place of uid) 99418_merge_retcode_ignored.patch (dispatch-conf ignored result of merge command) 100382_incomplete_world.patch (versioned atoms were not being added to world) broken-stat-perms.patch (result of conditional execution being unconditionally tested)
svn path=/main/branches/2.0/; revision=1960

+++ bin/ebuild.sh
@@ -867,15 +867,11 @@ dyn_compile() {
 		echo "!!! that you know what you are doing... You have 5 seconds to abort..."
 		echo
 
-		echo -ne "\a"; sleep 0.25 &>/dev/null; echo -ne "\a"; sleep 0.25 &>/dev/null
-		echo -ne "\a"; sleep 0.25 &>/dev/null; echo -ne "\a"; sleep 0.25 &>/dev/null
-		echo -ne "\a"; sleep 0.25 &>/dev/null; echo -ne "\a"; sleep 0.25 &>/dev/null
-		echo -ne "\a"; sleep 0.25 &>/dev/null; echo -ne "\a"; sleep 0.25 &>/dev/null
+		for x in 1 2 3 4 5 6 7 8; do
+			echo -ne "\a"
+			LC_ALL=C sleep 0.25
+		done
 
-		echo -ne "\a"; sleep 0,25 &>/dev/null; echo -ne "\a"; sleep 0,25 &>/dev/null
-		echo -ne "\a"; sleep 0,25 &>/dev/null; echo -ne "\a"; sleep 0,25 &>/dev/null
-		echo -ne "\a"; sleep 0,25 &>/dev/null; echo -ne "\a"; sleep 0,25 &>/dev/null
-		echo -ne "\a"; sleep 0,25 &>/dev/null; echo -ne "\a"; sleep 0,25 &>/dev/null
 		sleep 3
 	fi
 
@@ -1129,10 +1125,12 @@ dyn_install() {
 	local count=0
 	find "${D}/" -user  portage | while read file; do
 		count=$(( $count + 1 ))
-		[[ ! -L "${file}" ]] && s=$(stat_perms "$file")
-		if [ -z "${s}" ]; then
-			ewarn "failed stat_perm'ing $file.  User intervention during install isn't wise..."
-			continue
+		if [ ! -L "${file}" ]; then
+			s=$(stat_perms "$file")
+			if [ -z "${s}" ]; then
+				ewarn "failed stat_perm'ing $file.  User intervention during install isn't wise..."
+				continue
+			fi
 		fi
 		chown root "$file"
 		[[ ! -L "${file}" ]] && chmod "$s" "$file"
@@ -1144,10 +1142,12 @@ dyn_install() {
 	count=0
 	find "${D}/" -group portage | while read file; do
 		count=$(( $count + 1 ))
-		[[ ! -L "${file}" ]] && s=$(stat_perms "$file")
-		if [ -z "${s}" ]; then
-			echo "failed stat_perm'ing '$file' . User intervention during install isn't wise..."
-			continue
+		if [ ! -L "${file}" ]; then
+			s=$(stat_perms "$file")
+			if [ -z "${s}" ]; then
+				echo "failed stat_perm'ing '$file' . User intervention during install isn't wise..."
+				continue
+			fi
 		fi
 		chgrp 0 "${file}"
 		[[ ! -L "${file}" ]] && chmod "$s" "$file"
